Refactor artefact dialog file into smaller files
